Move `pants.backend.java.test` to `pants.jvm.test`.
[ci skip-rust]
