+++ pants.toml
@@ -176,7 +176,7 @@ java_parser = "src/python/pants/backend/java/dependency_inference/java_parser.lo
 scala_parser = "src/python/pants/backend/scala/dependency_inference/scala_parser.lockfile"
 
 [junit]
-lockfile = "src/python/pants/backend/java/test/junit.default.lockfile.txt"
+lockfile = "src/python/pants/jvm/junit.default.lockfile.txt"
 
 [google-java-format]
 lockfile = "src/python/pants/backend/java/lint/google_java_format/google_java_format.default.lockfile.txt"
