+++ src/python/pants/backend/experimental/java/register.py
@@ -14,13 +14,13 @@ from pants.backend.java.target_types import (
     JunitTestTarget,
 )
 from pants.backend.java.target_types import rules as target_types_rules
-from pants.backend.java.test import junit
 from pants.jvm import classpath, jdk_rules
 from pants.jvm import util_rules as jvm_util_rules
 from pants.jvm.dependency_inference import symbol_mapper
 from pants.jvm.goals import coursier
 from pants.jvm.resolve import coursier_fetch, coursier_setup, jvm_tool
 from pants.jvm.target_types import JvmArtifact
+from pants.jvm.test import junit
 
 
 def target_types():
