+++ src/python/pants/backend/experimental/scala/register.py
@@ -1,12 +1,7 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 from pants.backend.java.package import deploy_jar  # TODO: Should move to the JVM package.
-from pants.backend.java.target_types import (  # TODO: All of these should move to the JVM package.
-    DeployJar,
-    JunitTestsGeneratorTarget,
-    JunitTestTarget,
-)
-from pants.backend.java.test import junit  # TODO: Should move to the JVM package.
+from pants.backend.java.target_types import DeployJar  # TODO: Should move to the JVM package.
 from pants.backend.scala.compile import scalac
 from pants.backend.scala.dependency_inference import rules as dep_inf_rules
 from pants.backend.scala.goals import check, repl, tailor
@@ -22,13 +17,12 @@ from pants.jvm import util_rules as jvm_util_rules
 from pants.jvm.goals import coursier
 from pants.jvm.resolve import coursier_fetch, coursier_setup, jvm_tool
 from pants.jvm.target_types import JvmArtifact
+from pants.jvm.test import junit
 
 
 def target_types():
     return [
         DeployJar,
-        JunitTestTarget,
-        JunitTestsGeneratorTarget,
         JvmArtifact,
         ScalaJunitTestTarget,
         ScalaJunitTestsGeneratorTarget,
